refactor: reorganize CollectionSelect structure #31
